refactor: 💡 add package for custom hooks
