--- packages/charts/src/hooks/index.ts
@@ -1,3 +1,0 @@
-import { useTooltip } from './use-tooltip';
-
-export { useTooltip };
