+++ packages/react-hooks/README.md
@@ -1,0 +1,21 @@
+# `@keen.io/react-hooks`
+
+[![Known Vulnerabilities](https://snyk.io/test/github/keen/keen/badge.svg?targetFile=packages/react-hooks/package.json)](https://snyk.io/test/github/keen/keen?targetFile=packages/react-hooks/package.json)
+
+Set of reusable custom hooks for `React` ecosystem - used in `keen.io` projects.
+
+## Installing
+
+For Front-End applications use standard `npm` or [yarn](https://yarnpkg.com/lang/en/) package manager.
+
+```sh
+yarn install @keen.io/react-hooks
+```
+
+## Examples
+
+```js
+import { useLegend } from '@keen.io/react-hooks';
+
+const { disabledKeys, updateKeys } = useLegend();
+```
