+++ packages/react-hooks/jest.config.js
@@ -1,0 +1,9 @@
+module.exports = {
+  ...require('../../jest.config.base'),
+  setupFilesAfterEnv: ['<rootDir>/jest.setup.ts'],
+  moduleNameMapper: {
+    '^@keen.io/(.*)$': '<rootDir>/../$1/src',
+    '\\.(jpg|jpeg|png|gif|eot|otf|webp|svg|ttf|woff|woff2|mp4|webm|wav|mp3|m4a|aac|oga)$':
+      '<rootDir>/__mocks__/fileMock.js',
+  },
+};
