+++ packages/react-hooks/src/hooks/index.ts
@@ -1,0 +1,6 @@
+import useTooltip from './useTooltip';
+import useSlider from './useSlider';
+import useLegend from './useLegend';
+import useSearch from './useSearch';
+
+export { useTooltip, useSlider, useLegend, useSearch };
