+++ packages/react-hooks/src/hooks/useLegend/index.ts
@@ -1,0 +1,3 @@
+import { useLegend } from './useLegend';
+
+export default useLegend;
