+++ packages/react-hooks/src/hooks/useSearch/index.ts
@@ -1,0 +1,3 @@
+import { useSearch } from './useSearch';
+
+export default useSearch;
