+++ packages/react-hooks/src/hooks/useSlider/index.ts
@@ -1,0 +1,3 @@
+import { useSlider } from './useSlider';
+
+export default useSlider;
