+++ packages/react-hooks/src/hooks/useSlider/useSlider.ts
