+++ packages/react-hooks/src/hooks/useTooltip/index.ts
@@ -1,0 +1,3 @@
+import { useTooltip } from './useTooltip';
+
+export default useTooltip;
