+++ packages/react-hooks/src/hooks/useTooltip/types.ts
@@ -1,0 +1,8 @@
+export type DataSelector = (number | string)[];
+
+export type TooltipState = {
+  visible: boolean;
+  selectors: { selector: DataSelector; color: string }[];
+  x: number;
+  y: number;
+};
