+++ packages/react-hooks/src/index.ts
@@ -1,0 +1,3 @@
+import { useTooltip, useSlider, useLegend, useSearch } from './hooks';
+
+export { useTooltip, useSlider, useLegend, useSearch };
