--- packages/ui-core/src/hooks/index.ts
@@ -1,3 +1,0 @@
-import { useSearch } from './use-search';
-
-export { useSearch };
