+++ packages/widgets/src/components/area/area-chart.widget.tsx
@@ -8,6 +8,7 @@ import {
   theme as defaultTheme,
   sortAreaKeys,
 } from '@keen.io/charts';
+import { useLegend } from '@keen.io/react-hooks';
 
 import ChartWidget from '../chart-widget.component';
 import WidgetHeading from '../widget-heading.component';
@@ -16,8 +17,6 @@ import {
   LegendSocket,
   TitleSocket,
 } from '../widget-sockets.component';
-
-import { useLegend } from '../../hooks';
 
 import { legendSettings } from '../../widget-settings';
 import { WidgetSettings, LegendSettings } from '../../types';
