--- packages/widgets/src/hooks/index.ts
@@ -1,4 +1,0 @@
-import { useLegend } from './use-legend.hook';
-import { useSlider } from './use-slider.hook';
-
-export { useLegend, useSlider };
